# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ConfirmedExpenseParticipant:
    properties:
      fees:
        items:
          $ref: '#/definitions/dto.ConfirmedItemShare'
        type: array
      feesTotal:
        type: number
      hasProxy:
        type: boolean
      items:
        items:
          $ref: '#/definitions/dto.ConfirmedItemShare'
        type: array
      itemsTotal:
        type: number
      profile:
        $ref: '#/definitions/dto.SimpleProfile'
      proxyProfile:
        $ref: '#/definitions/dto.SimpleProfile'
      total:
        type: number
    type: object
  dto.ConfirmedItemShare:
    properties:
      baseAmount:
        type: number
      id:
        type: string
      name:
        type: string
      shareAmount:
        type: number
      shareRate:
        type: number
    type: object
...
  dto.ExpenseBillResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      imageUrl:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
  dto.ExpenseConfirmationResponse:
    properties:
      currency:
        type: string
      description:
        type: string
      id:
        type: string
      participants:
        items:
          $ref: '#/definitions/dto.ConfirmedExpenseParticipant'
        type: array
      payer:
        $ref: '#/definitions/dto.SimpleProfile'
      totalAmount:
        type: number
    type: object
  dto.ExpenseItemResponse:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      groupExpenseId:
        type: string
      id:
        type: string
      name:
        type: string
      participants:
        items:
          $ref: '#/definitions/dto.ItemParticipantResponse'
        type: array
      quantity:
        type: integer
      updatedAt:
        type: string
    type: object
  dto.ExpenseParticipantResponse:
    properties:
      hasProxy:
        type: boolean
      participantProfile:
        $ref: '#/definitions/dto.SimpleProfile'
      proxyProfile:
        $ref: '#/definitions/dto.SimpleProfile'
      shareAmount:
        type: number
    type: object
  dto.ExpenseParticipantsRequest:
    properties:
      participantProfileIds:
        items:
          type: string
        minItems: 1
        type: array
      payerProfileId:
        type: string
      proxyByProfileIds:
        additionalProperties:
          type: string
        type: object
    required:
    - participantProfileIds
    - payerProfileId
    type: object
...
  dto.FeeParticipantResponse:
    properties:
      profile:
        $ref: '#/definitions/dto.SimpleProfile'
      shareAmount:
        type: number
    type: object
...
  dto.GroupExpenseResponse:
    properties:
      bill:
        $ref: '#/definitions/dto.ExpenseBillResponse'
      billExists:
        type: boolean
      confirmationPreview:
        $ref: '#/definitions/dto.ExpenseConfirmationResponse'
      createdAt:
        type: string
      creator:
        $ref: '#/definitions/dto.SimpleProfile'
      currency:
        type: string
      description:
        type: string
      feesTotalAmount:
        type: number
      id:
        type: string
      isPreviewable:
        type: boolean
      items:
        items:
          $ref: '#/definitions/dto.ExpenseItemResponse'
        type: array
      itemsTotalAmount:
        type: number
      otherFees:
        items:
          $ref: '#/definitions/dto.OtherFeeResponse'
        type: array
      participants:
        items:
          $ref: '#/definitions/dto.ExpenseParticipantResponse'
        type: array
      payer:
        allOf:
        - $ref: '#/definitions/dto.SimpleProfile'
        description: Relationships
      status:
        type: string
      totalAmount:
        type: number
      updatedAt:
        type: string
    type: object
...
  dto.ItemParticipantResponse:
    properties:
      allocatedAmount:
        type: number
      profile:
        $ref: '#/definitions/dto.SimpleProfile'
      weight:
        type: integer
    type: object
...
  dto.NewDraftRequest:
    properties:
      description:
        type: string
    type: object
...
  dto.NotificationResponse:
    properties:
      createdAt:
        type: string
      entityId:
        type: string
      entityType:
        type: string
      id:
        type: string
      metadata:
        type: object
      readAt:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  dto.OtherFeeResponse:
    properties:
      amount:
        type: number
      calculationMethod:
        type: string
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      participants:
        items:
          $ref: '#/definitions/dto.FeeParticipantResponse'
        type: array
      updatedAt:
        type: string
    type: object
...
  dto.SimpleProfile:
    properties:
      avatar:
        type: string
      id:
        type: string
      isUser:
        type: boolean
      name:
        type: string
    type: object
...
  response.JSONResponse-array_dto_GroupExpenseResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.GroupExpenseResponse'
        type: array
      errors:
        items: {}
        type: array
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
  response.JSONResponse-array_dto_NotificationResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.NotificationResponse'
        type: array
      errors:
        items: {}
        type: array
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
...
  response.JSONResponse-dto_ExpenseConfirmationResponse:
    properties:
      data:
        $ref: '#/definitions/dto.ExpenseConfirmationResponse'
      errors:
        items: {}
        type: array
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
...
  response.JSONResponse-dto_GroupExpenseResponse:
    properties:
      data:
        $ref: '#/definitions/dto.GroupExpenseResponse'
      errors:
        items: {}
        type: array
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
...
  response.Pagination:
    properties:
      currentPage:
        type: integer
      hasNextPage:
        type: boolean
      hasPrevPage:
        type: boolean
      totalData:
        type: integer
      totalPages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Cashback backend API
  title: Cashback API
  version: "1.0"
paths:
...
  /group-expenses:
    get:
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: 'Filter by ownership (default: owned)'
        in: query
        name: ownership
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-array_dto_GroupExpenseResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all group expenses
      tags:
      - group-expenses
    post:
      consumes:
      - application/json
      parameters:
      - description: New draft payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.NewDraftRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.JSONResponse-dto_GroupExpenseResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a draft group expense
      tags:
      - group-expenses
  /group-expenses/{groupExpenseId}:
    delete:
      parameters:
      - description: Group expense ID
        in: path
        name: groupExpenseId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a group expense
      tags:
      - group-expenses
    get:
      parameters:
      - description: Group expense ID
        in: path
        name: groupExpenseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-dto_GroupExpenseResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get group expense details
      tags:
      - group-expenses
...
  /group-expenses/{groupExpenseId}/confirmed:
    patch:
      parameters:
      - description: Group expense ID
        in: path
        name: groupExpenseId
        required: true
        type: string
      - description: Set to true for dry run
        in: query
        name: dry-run
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-dto_ExpenseConfirmationResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Confirm a draft group expense
      tags:
      - group-expenses
...
  /group-expenses/{groupExpenseId}/participants:
    put:
      consumes:
      - application/json
      parameters:
      - description: Group expense ID
        in: path
        name: groupExpenseId
        required: true
        type: string
      - description: Participants payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ExpenseParticipantsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Sync participants of a group expense
      tags:
      - group-expenses
...
  /group-expenses/recent:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-array_dto_GroupExpenseResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get recent group expenses
      tags:
      - group-expenses
  /notifications:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-array_dto_NotificationResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get unread notifications
      tags:
      - notifications
    patch:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - notifications
  /notifications/{notificationId}:
    patch:
      parameters:
      - description: Notification ID
        in: path
        name: notificationId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark a notification as read
      tags:
      - notifications
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"